MERDAAAAAAA

+++ primos.h
@@ -7,7 +7,7 @@
 #include <chrono>
 #include "potenciaModular.h"
 
-int gerarPossivelPrimo()
+long long gerarPossivelPrimo()
 {
 
     // Pega a seed a partir do relógid do sistema
@@ -24,11 +24,11 @@ int gerarPossivelPrimo()
     return possivelPrimo;
 }
 
-long long gerarPossivelPrimoIntervalo(long long vMax)
+unsigned long long gerarPossivelPrimoIntervalo(unsigned long long vMax)
 {
     // Define um valor mínimo para não correr o risco de gerar valores muito pequenos.
     int minBits = 18;
-    long long m = pow(2, minBits);
+    unsigned long long m = pow(2, minBits);
 
     // Pega a seed a partir do relógid do sistema
     unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
@@ -36,7 +36,7 @@ long long gerarPossivelPrimoIntervalo(long long vMax)
     // Seta a seed para o gerador de números aleatórios de 24 bits
     std::ranlux24 generator(seed);
 
-    long long num;
+    unsigned long long num;
     do
     {
         num = generator();
@@ -55,7 +55,7 @@ long long gerarPossivelPrimoIntervalo(long long vMax)
     a -> Números primos
     a ^ p - 1 != 1 (mod p) -> Tem certeza que não é primo
 */
-bool verificaPrimo(long long possivelPrimo) // Implementação do teste do pequeno teorema de Fermat
+bool verificaPrimo(unsigned long long possivelPrimo) // Implementação do teste do pequeno teorema de Fermat
 {
     int primos[] = {2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53,
                     59, 61, 67, 71, 73, 79, 83, 89, 97, 101, 103, 107, 109, 113, 127, 131, 137, 139, 149, 151, 157,
@@ -65,9 +65,9 @@ bool verificaPrimo(long long possivelPrimo) // Implementação do teste do peque
                     499, 503, 509, 521, 523, 541, 547, 557, 563, 569, 571, 577, 587, 593, 599, 601, 607, 613, 617,
                     619, 631, 641, 643, 647, 653, 659, 661, 673, 677, 683, 701, 709, 719, 727, 733, 739, 743, 751};
 
-    long long size = sizeof(primos) / sizeof(int);
+    unsigned long long size = sizeof(primos) / sizeof(int);
 
-    long long a;
+    unsigned long long a;
 
     for (int i = 0; i < size; i++)
     {
@@ -80,7 +80,7 @@ bool verificaPrimo(long long possivelPrimo) // Implementação do teste do peque
     return true;
 }
 
-long long gerarPseudoPrimo()
+unsigned long long gerarPseudoPrimo()
 {
     bool verifica = false;
